Aggregate file and user logic
It seems redundant to have separate logic for dealing with files and users, e.g. calling the 'files' endpoint to obtain a list of files and then the 'users' endpoint to obtain a list of users that can view one of those files. As a result, I aggregated the controllers and services to the extent that user logic was left with user registration only.